Fixed Hough accumulate CUDA kernel - was indexing mask by (x, y), but OpenCV uses (y, x), essentially flipping the image along the diagonal

+++ ProblemSets/ps1_cpp/src/Config.h
@@ -55,7 +55,7 @@ public:
     };
 
     struct Hough : BasicConfig {
-        size_t _rhoBinSize, _thetaBinSize, _numPeaks;
+        float _rhoBinSize, _thetaBinSize, _numPeaks;
         int _threshold;
 
         Hough() = default;
